Adiciona funcionalidade de busca de produtos e refatora a estrutura de dados para suportar pesquisa

+++ src/helpers/mod.rs
@@ -1,4 +1,4 @@
-pub fn format_int_to_decimal(valor: i32) -> String {
+pub fn format_int_to_decimal(valor: i64) -> String {
     let part_int = valor / 100;
     let part_float = (valor % 100).abs();
     format!("R$ {},{:02}", part_int, part_float)
